+++ src/pages/board.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 
-const Board = ({ id, board }) => {
+const Board = ({ match }) => {
+  const board = Number(match.params.board);
+  const boardId = Number(match.params.boardId);
   return (
     <>
       <h1>{board.title}</h1>
